refactor(ErDiagram): extract layout hook into own file
