+++ src/app/components/ErDiagram/ErDiagram.tsx
@@ -1,5 +1,4 @@
-import ELK, { ElkNode } from "elkjs/lib/elk.bundled.js";
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 import ReactFlow, {
   Background,
   BackgroundVariant,
@@ -24,116 +23,7 @@ import { FaWandMagicSparkles } from "react-icons/fa6";
 import { Tooltip } from "@chakra-ui/react";
 import { useTranslations } from "next-intl";
 import { ErNotation } from "../../types/ErNotation";
-
-const elk = new ELK();
-
-const useLayoutedElements = () => {
-  const { getNodes, setNodes, getEdges, fitView } = useReactFlow();
-  const defaultOptions = {
-    "elk.algorithm": "layered",
-    // "elk.layered.spacing.nodeNodeBetweenLayers": 50,
-    // "elk.spacing.nodeNode": 100000,
-  };
-
-  const getLayoutedElements = useCallback(
-    async (options: { [key: string]: string }) => {
-      const layoutOptions = { ...defaultOptions, ...options };
-      const flowNodes = getNodes();
-      const flowEdges = getEdges();
-
-      const aggregationNodeIds = flowNodes
-        .filter((n) => n.type === "aggregation")
-        .map((n) => n.id);
-
-      // aggregations are treated as an ELK subgraph
-      const aggregationGraphs = aggregationNodeIds.map((aggNodeId) => {
-        const childrenNodes = flowNodes.filter(
-          (n) => n.parentNode === aggNodeId,
-        );
-        const childrenEdges = flowEdges.filter((e) => {
-          return (
-            childrenNodes.some((n) => n.id === e.source) &&
-            childrenNodes.some((n) => n.id === e.target)
-          );
-        });
-
-        return {
-          id: aggNodeId,
-          layoutOptions: layoutOptions,
-          children: childrenNodes,
-          edges: childrenEdges,
-        };
-      });
-
-      const notInAggregationNodes = flowNodes.filter(
-        (n) =>
-          // is not a child of an aggregation subgraph
-          aggregationGraphs.every((agg) =>
-            agg.children.every((c) => c.id !== n.id),
-          ) &&
-          // is not an aggregation node
-          n.type !== "aggregation",
-      );
-
-      const notInAggregationEdges = flowEdges.filter((edge) =>
-        aggregationGraphs.every((agg) =>
-          agg.edges.every((aggEd) => aggEd.id !== edge.id),
-        ),
-      );
-
-      const graph = {
-        id: "root",
-        layoutOptions: layoutOptions,
-        children: [...notInAggregationNodes, ...aggregationGraphs],
-        edges: notInAggregationEdges,
-      };
-
-      const { children } = await elk.layout(graph as unknown as ElkNode);
-      const layoutedNodes: Node[] = [];
-
-      children?.forEach((node) => {
-        if (node.hasOwnProperty("children")) {
-          // mutate the aggregation subgraph back into a regular node
-          const originalAgg = flowNodes.find((fn) => fn.id === node.id);
-          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-          // @ts-ignore
-          node = {
-            ...originalAgg,
-            ...node,
-          };
-          node.data.width = node.width;
-          node.data.height = node.height;
-
-          layoutedNodes.push({
-            ...node,
-            position: { x: node.x, y: node.y },
-          } as Node);
-          // and bring the aggregation child nodes to the main graph
-          node.children?.forEach((childNode) => {
-            layoutedNodes.push({
-              ...childNode,
-              position: { x: childNode.x!, y: childNode.y! },
-            } as Node);
-          });
-        } else {
-          layoutedNodes.push({
-            ...node,
-            position: { x: node.x!, y: node.y! },
-          } as Node);
-        }
-      });
-      console.log(layoutedNodes);
-
-      setNodes(layoutedNodes as Node[]);
-      window.requestAnimationFrame(() => {
-        fitView();
-      });
-    },
-    [],
-  );
-
-  return { getLayoutedElements };
-};
+import useLayoutedElements from "./useLayoutedElements";
 
 type ErDiagramProps = {
   erDoc: ER;
